+++ docs/source/api_ref_training.rst
@@ -1,0 +1,18 @@
+==================
+torchtune.training
+==================
+
+.. currentmodule:: torchtune.training
+
+Reduced Precision
+------------------
+
+Utilities for working in a reduced precision setting.
+
+.. autosummary::
+    :toctree: generated/
+    :nosignatures:
+
+    get_dtype
+    set_default_dtype
+    validate_expected_param_dtype
