+++ recipes/dev/lora_finetune_fsdp2.py
@@ -24,7 +24,7 @@ from torch.distributed.algorithms._checkpoint.checkpoint_wrapper import (
 
 from torch.optim import Optimizer
 from torch.utils.data import DataLoader, DistributedSampler
-from torchtune import config, modules, utils
+from torchtune import config, modules, training, utils
 from torchtune.data import padded_collate
 from torchtune.datasets import ConcatDataset
 from torchtune.modules.peft import (
@@ -111,7 +111,7 @@ class LoRAFinetuneRecipeDistributed(FTRecipeInterface):
             raise RuntimeError("FSDP2 recipe is only available on PyTorch nightlies")
 
         self._device = utils.get_device(device=cfg.device)
-        self._dtype = utils.get_dtype(cfg.dtype, device=self._device)
+        self._dtype = training.get_dtype(cfg.dtype, device=self._device)
 
         if self._dtype == torch.float16:
             raise ValueError(
@@ -299,7 +299,7 @@ class LoRAFinetuneRecipeDistributed(FTRecipeInterface):
             )
             init_start = time.perf_counter()
 
-        with utils.set_default_dtype(self._dtype), torch.device("meta"):
+        with training.set_default_dtype(self._dtype), torch.device("meta"):
             model = config.instantiate(cfg_model)
 
         self.adapter_params = get_adapter_params(model)
@@ -344,7 +344,7 @@ class LoRAFinetuneRecipeDistributed(FTRecipeInterface):
         else:
             lora_missing, lora_unexpected = None, None
 
-        with utils.set_default_dtype(self._dtype), self._device:
+        with training.set_default_dtype(self._dtype), self._device:
             lora_device = "cpu" if cfg_fsdp and cfg_fsdp.cpu_offload else self._device
             for m in model.modules():
                 if isinstance(m, LoRALinear) and not lora_weights_state_dict:
