+++ recipes/lora_dpo_distributed.py
@@ -24,7 +24,7 @@ from torch.distributed.fsdp import (
 )
 from torch.optim import Optimizer
 from torch.utils.data import DataLoader, DistributedSampler
-from torchtune import config, modules, utils
+from torchtune import config, modules, training, utils
 from torchtune.data import CROSS_ENTROPY_IGNORE_IDX
 from torchtune.datasets import ConcatDataset
 from torchtune.modules import rlhf
@@ -117,7 +117,7 @@ class LoRADPORecipeDistributed(FTRecipeInterface):
 
     def __init__(self, cfg: DictConfig) -> None:
         self._device = utils.get_device(device=cfg.device)
-        self._dtype = utils.get_dtype(cfg.dtype, device=self._device)
+        self._dtype = training.get_dtype(cfg.dtype, device=self._device)
 
         if self._dtype == torch.float16:
             raise ValueError(
@@ -310,7 +310,7 @@ class LoRADPORecipeDistributed(FTRecipeInterface):
             log.info("FSDP is enabled. Instantiating Model on CPU for Rank 0 ...")
             init_start = time.perf_counter()
 
-            with utils.set_default_dtype(self._dtype):
+            with training.set_default_dtype(self._dtype):
                 model = config.instantiate(cfg_model)
 
             log.info(
@@ -343,7 +343,7 @@ class LoRADPORecipeDistributed(FTRecipeInterface):
 
         else:
             # For non-zero ranks, load the model on meta device
-            with utils.set_default_dtype(self._dtype), torch.device("meta"):
+            with training.set_default_dtype(self._dtype), torch.device("meta"):
                 model = config.instantiate(cfg_model)
 
         if self._dtype == torch.bfloat16:
