+++ tests/torchtune/models/phi3/test_lora_phi3.py
@@ -12,7 +12,7 @@ import torch
 from tests.test_utils import assert_expected, fixed_init_model
 from torch import nn
 from torchao.dtypes.nf4tensor import NF4Tensor
-from torchtune import utils
+from torchtune import training
 from torchtune.models.phi3 import lora_phi3, phi3
 from torchtune.models.phi3._component_builders import lora_phi3_self_attention
 from torchtune.modules.peft import get_merged_lora_ckpt, LoRALinear
@@ -205,7 +205,7 @@ class TestLoRAPhi3:
 
     @pytest.mark.parametrize("dtype", [torch.bfloat16, torch.float32])
     def test_qlora_phi3_parity(self, dtype, inputs):
-        with utils.set_default_dtype(dtype):
+        with training.set_default_dtype(dtype):
             model_ref = self.get_lora_phi3(
                 lora_modules=["q_proj", "v_proj", "k_proj", "output_proj"],
                 apply_lora_to_mlp=True,
@@ -235,7 +235,7 @@ class TestLoRAPhi3:
 
     @pytest.mark.parametrize("dtype", [torch.bfloat16, torch.float32])
     def test_qlora_phi3_state_dict(self, dtype):
-        with utils.set_default_dtype(dtype):
+        with training.set_default_dtype(dtype):
             model_ref = self.get_lora_phi3(
                 lora_modules=["q_proj", "v_proj", "k_proj", "output_proj"],
                 apply_lora_to_mlp=True,
@@ -271,7 +271,7 @@ class TestLoRAPhi3:
 
     @pytest.mark.parametrize("dtype", [torch.bfloat16, torch.float32])
     def test_qlora_phi3_merged_state_dict(self, dtype):
-        with utils.set_default_dtype(dtype):
+        with training.set_default_dtype(dtype):
             qlora = self.get_lora_phi3(
                 lora_modules=["q_proj", "v_proj", "k_proj", "output_proj"],
                 apply_lora_to_mlp=True,
@@ -295,7 +295,7 @@ class TestLoRAPhi3:
                 assert v.dtype == dtype
 
         # Ensure checkpoint can be loaded into non-LoRA model
-        with utils.set_default_dtype(dtype):
+        with training.set_default_dtype(dtype):
             phi3 = self.get_ref_phi3(vocab_size=50, embed_dim=512)
 
         phi3.load_state_dict(merged_ckpt)
