+++ tests/torchtune/modules/peft/test_lora.py
@@ -12,7 +12,7 @@ import torch
 from tests.test_utils import fixed_init_model
 from torch import nn
 from torchao.dtypes.nf4tensor import NF4Tensor, to_nf4
-from torchtune import utils
+from torchtune import training
 from torchtune.modules.common_utils import reparametrize_as_dtype_state_dict_post_hook
 from torchtune.modules.peft import LoRALinear
 from torchtune.utils.seed import set_seed
@@ -62,7 +62,7 @@ class TestLoRALinear:
 
     @pytest.fixture
     def qlora_linear(self, in_dim, out_dim) -> LoRALinear:
-        with utils.set_default_dtype(torch.bfloat16):
+        with training.set_default_dtype(torch.bfloat16):
             qlora_linear = LoRALinear(
                 in_dim=512,
                 out_dim=512,
@@ -113,7 +113,7 @@ class TestLoRALinear:
 
     @pytest.mark.parametrize("dtype", [torch.bfloat16, torch.float32])
     def test_qlora_parity(self, dtype):
-        with utils.set_default_dtype(dtype):
+        with training.set_default_dtype(dtype):
             qlora_linear = LoRALinear(
                 in_dim=512,
                 out_dim=512,
@@ -145,7 +145,7 @@ class TestLoRALinear:
 
     @pytest.mark.parametrize("dtype", [torch.bfloat16, torch.float32])
     def test_quantized_state_dict(self, dtype):
-        with utils.set_default_dtype(dtype):
+        with training.set_default_dtype(dtype):
             lora_linear = LoRALinear(
                 in_dim=512,
                 out_dim=512,
