+++ torchtune/training/__init__.py
@@ -1,0 +1,17 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+from torchtune.training.precision import (
+    get_dtype,
+    set_default_dtype,
+    validate_expected_param_dtype,
+)
+
+__all__ = [
+    "get_dtype",
+    "set_default_dtype",
+    "validate_expected_param_dtype",
+]
